Rating Modification

+++ database/migrations/2020_09_30_141127_courses.php
@@ -37,7 +37,7 @@ class Courses extends Migration {
             $table->double('discount_percentage')->default(0);
             $table->enum('discount_amount_type', ['cash', 'percentage'])->default('percentage');
 
-            $table->integer('rating')->default(0);
+            $table->double('rating')->default(0);
             $table->integer('rating_count')->default(0);
 
             $table->string('course_image')->default('default_course.jpg');
